Organize project under zpl_to_pdf package
